Consolidate otel queries and remove dead code
