+++ frontend/src/pages/RequestDetailsPage/v2/SummaryV2.tsx
@@ -7,7 +7,6 @@ import {
   SEMATTRS_EXCEPTION_TYPE,
 } from "@opentelemetry/semantic-conventions";
 import { useMemo } from "react";
-import { TextOrJsonViewer } from "../TextJsonViewer";
 import { FpxCard, RequestMethod } from "../shared";
 import { BodyViewerV2 } from "./BodyViewerV2";
 import {
