+++ frontend/src/queries/traces-otel.ts
@@ -101,3 +101,38 @@ async function fetchOtelTrace(context: QueryFunctionContext<[string, string]>) {
     })
     .then((data) => SpansSchema.parse(data));
 }
+
+export function useOtelTraces() {
+  return useQuery({
+    queryKey: ["otel-traces"],
+    queryFn: (): Promise<OtelSpan[]> => {
+      return fetch("/v1/traces")
+        .then((res) => res.json())
+        .then((r) => {
+          // Uncomment the following line for inspection:
+          // console.log("Otel Traces before decoding:", r);
+          return r;
+        })
+        .then((r) =>
+          r.map((r: { parsedPayload: unknown; rawPayload: unknown }) =>
+            toOtelSpan(r?.parsedPayload, r?.rawPayload),
+          ),
+        );
+    },
+  });
+}
+
+function toOtelSpan(
+  t: unknown,
+  rawPayload: unknown,
+): (OtelSpan & { rawPayload: unknown }) | null {
+  const result = OtelSpanSchema.safeParse(t);
+  if (!result.success) {
+    console.error("OtelSpanSchema parse error:", result.error.format());
+    return null;
+  }
+  return {
+    ...result.data,
+    rawPayload,
+  };
+}
